修改了数据类型

+++ src/main/java/com/baoquan/shuxin/service/spi/news/NewsService.java
@@ -32,7 +32,7 @@ public interface NewsService {
     * @return
     * @throws
     */
-    News updateNews(Integer id);
+    News updateNews(Long id);
 
     /**
      * 修改新闻和添加新闻
